djangoProject/requirements.txt
Django~=3.1
djangorestframework~=3.12
psycopg2
gunicorn
gevent
Pillow
pytz
django-cors-headers
django-filter
